+++ resources/components/laraknife/buttons/button-position.blade.php
@@ -1,4 +1,4 @@
-@props(['position' => '', 'name' => 'btnOK', 'label' => 'OK', 'width1' => 2, 'width2' => 4])
+@props(['position' => '', 'name' => 'btnOK', 'label' => 'OK', 'width1' => 2, 'width2' => 4, 'class' => ''])
 @if($position === 'alone' || $position === 'first')
 <div class="row">
 @endif
@@ -6,7 +6,7 @@
 <x-laraknife.layout.col-empty width="{{$width1}}"/>
 @endif
 <div class="col-md-{{ $width2 }}">
-    <button class="lkn-button expand100" name="btnSubmit" value="{{$name}}">{{ __($label) }}</button>
+    <button class="lkn-button expand100 {{ $class }}" name="btnSubmit" value="{{$name}}">{{ __($label) }}</button>
 </div>
 @if($position === 'alone' || $position === 'last')
 </div>