+++ resources/components/laraknife/panels/main-header.blade.php
@@ -1,2 +1,2 @@
 @props(['title' => ''])
-<h1 class="text-center lkn-expand100">{{$title}}</h2>
+<h1 class="text-center lkn-expand100">{{$title}}</h1>