+++ resources/helpers/StringHelper.php
@@ -42,18 +42,60 @@ class StringHelper
         return $rc;
     }
     /**
-     * Converts a string into a word usable as part of an URL.
-     * @param string $text the text to convert
-     * @param NULL|int $maxLength the result is never longer than that
-     * @return string a word usable as part of the string
+     * Converts a string into an assoziative array. Inversion of implodeAssoc().
+     * @param string $source the string to convert
+     * @param string $separatorLine the separator between the pair entries
+     * @param string $separatorPair the separator between key and value
+     * @param array $array the array to convert
      */
-    public static function textToUrl(string $text, ?int $maxLength = null): string
+    public static function explodeAssoc(string $source, string $separatorLine = "\n", string $separatorPair = '='): array
     {
-        $rc = preg_replace(['/[^\w.+-]+/', '/__+/'], ['_', '_'], $text);
-        if ($maxLength != null && strlen($rc) > $maxLength) {
-            $rc = substr($rc, $maxLength);
+        $rc = [];
+        $pos = 0;
+        while (($end = strpos($source, $separatorLine, $pos)) !== false) {
+            if (($keyEnd = strpos($source, $separatorPair, $pos)) === false) {
+                $rc[substr($source, $pos, $end)] = '';
+            } else {
+                $key = substr($source, $pos, $keyEnd - $pos);
+                $value = substr($source, $keyEnd + 1, $keyEnd - $pos - 1);
+                $rc[$key] = $value;
+            }
+            $pos = $end + 1;
         }
-        return strtolower($rc);
+        if ($pos < strlen($source) - 1) {
+            if (($keyEnd = strpos($source, $separatorPair, $pos)) === false) {
+                $rc[substr($source, $pos, $end)] = '';
+            } else {
+                $key = substr($source, $pos, $keyEnd - $pos);
+                $value = substr($source, $keyEnd + 1, $keyEnd - $pos - 1);
+                $rc[$key] = $value;
+            }
+        }
+        return $rc;
+    }
+
+    /**
+     * Converts an assoziative array into a string.
+     * @param array $array the array to convert
+     * @param string $separatorLine the separator between the pair entries
+     * @param string $separatorPair the separator between key and value
+     * @return string the serialized array
+     */
+    public static function implodeAssoc(array &$array, string $separatorLine = "\n", string $separatorPair = '=', bool $ignoreMeta = true): string
+    {
+        $rc = '';
+        foreach ($array as $key => $value) {
+            if ($ignoreMeta && ($key[0] === '_' || str_starts_with($key, 'btn'))) {
+                continue;
+            } else {
+                if ($rc === '') {
+                    $rc = "$key$separatorPair$value";
+                } else {
+                    $rc .= "$separatorLine$key$separatorPair$value";
+                }
+            }
+        }
+        return $rc;
     }
     public static function randomChar(string $charSet): string
     {
@@ -64,7 +106,7 @@ class StringHelper
     {
         $rc = '';
         $max = strlen($charSet) - 1;
-        for ($ix = 0; $ix < $length; $ix++){
+        for ($ix = 0; $ix < $length; $ix++) {
             $rc .= $charSet[rand(0, $max)];
         }
         return $rc;
@@ -87,6 +129,20 @@ class StringHelper
             $rc = $word;
         }
         return $rc;
+    }
+    /**
+     * Converts a string into a word usable as part of an URL.
+     * @param string $text the text to convert
+     * @param NULL|int $maxLength the result is never longer than that
+     * @return string a word usable as part of the string
+     */
+    public static function textToUrl(string $text, ?int $maxLength = null): string
+    {
+        $rc = preg_replace(['/[^\w.+-]+/', '/__+/'], ['_', '_'], $text);
+        if ($maxLength != null && strlen($rc) > $maxLength) {
+            $rc = substr($rc, $maxLength);
+        }
+        return strtolower($rc);
     }
     /**
      * Converts the $string into a capital string: the first character will be uppercase.