+++ resources/views/note/edit.blade.php
@@ -13,6 +13,8 @@
                     value="{{ $context->valueOf('title') }}" width2="4" />
                 <x-laraknife.forms.combobox position="last" name="visibility_scope" label="Visibility" :options="$optionsVisibility"
                     width2="4" />
+                <x-laraknife.forms.combobox position="alone" name="group_id" label="Group" :options="$optionsGroup"
+                    width2="4" />
                 <x-laraknife.forms.text position="alone" name="body" label="Body"
                     value="{{ $context->valueOf('body') }}" width2="10" rows="10" />
             </x-laraknife.layout.nav-tabs>